NATO Member Defense Expenditures

A sample dataset of NATO/OTAN member defense expenditures.

nato

Format

A data.frame of 232 rows and 14 variables.

Variables:

country_stanag

Country code in NATO STANAG format

year

Year, from 2012 to 2019. 2018-2019 numbers may be estimates.

Defense expenditure (USD, current prices)

Defense expenditures in US dollars, using current prices and exchange rates.

Defense expenditure (USD, 2015 prices)

Defense expenditures in US dollars, using 2015 prices and exchange rates.

Defense expenditure (% real GDP)

Defense expenditure as a percentage of real gross domestic product. Based on 2015 prices.

Defense expenditure annual real change (% GDP)

Annual change in defense expenditure as a percentage of real gross domestic product. Based on 2015 prices.

Real GDP (2015 prices)

Real gross domestic product in 2015 US dollars and at 2015 exchange rates.

GDP per capita (USD)

Gross domestic product per capita in 2015 US dollars and at 2015 exchange rates.

Defense expenditure per capita (USD)

Defense expenditure per capita in 2015 US dollars.

Military personnel

Number of military personnel

Equipment expenditure (%)

Percent of defense expenditure spent on equipment. Includes major equipment expenditure and R&D devoted to major equipment.

Personnel expenditure (%)

Percentage of defense expenditure spent on personnel. Includes both military and civilian expenditure and pensions.

Infrastructure expenditure (%)

Percentage of defense expenditure spent on infrastructure. Includes NATO common infrastructure and national military construction.

Other expenditure (%)

Percentage of defense expenditure spent on other categories besides equipment, personnel, and infrastructure. Includes operations and maintenance expenditure, other R&D expenditure, and other expenditure not otherwise captured.

Source

https://www.nato.int/cps/en/natohq/news_167080.htm

Examples

as_country_name(nato$country_stanag, from = 'stanag')

#>   [1] "Albania"     "Albania"     "Albania"     "Albania"     "Albania"    
#>   [6] "Albania"     "Albania"     "Albania"     "Belgium"     "Belgium"    
#>  [11] "Belgium"     "Belgium"     "Belgium"     "Belgium"     "Belgium"    
#>  [16] "Belgium"     "Bulgaria"    "Bulgaria"    "Bulgaria"    "Bulgaria"   
#>  [21] "Bulgaria"    "Bulgaria"    "Bulgaria"    "Bulgaria"    "Canada"     
#>  [26] "Canada"      "Canada"      "Canada"      "Canada"      "Canada"     
#>  [31] "Canada"      "Canada"      "Czechia"     "Czechia"     "Czechia"    
#>  [36] "Czechia"     "Czechia"     "Czechia"     "Czechia"     "Czechia"    
#>  [41] "Germany"     "Germany"     "Germany"     "Germany"     "Germany"    
#>  [46] "Germany"     "Germany"     "Germany"     "Denmark"     "Denmark"    
#>  [51] "Denmark"     "Denmark"     "Denmark"     "Denmark"     "Denmark"    
#>  [56] "Denmark"     "Spain"       "Spain"       "Spain"       "Spain"      
#>  [61] "Spain"       "Spain"       "Spain"       "Spain"       "Estonia"    
#>  [66] "Estonia"     "Estonia"     "Estonia"     "Estonia"     "Estonia"    
#>  [71] "Estonia"     "Estonia"     "France"      "France"      "France"     
#>  [76] "France"      "France"      "France"      "France"      "France"     
#>  [81] "UK"          "UK"          "UK"          "UK"          "UK"         
#>  [86] "UK"          "UK"          "UK"          "Greece"      "Greece"     
#>  [91] "Greece"      "Greece"      "Greece"      "Greece"      "Greece"     
#>  [96] "Greece"      "Croatia"     "Croatia"     "Croatia"     "Croatia"    
#> [101] "Croatia"     "Croatia"     "Croatia"     "Croatia"     "Hungary"    
#> [106] "Hungary"     "Hungary"     "Hungary"     "Hungary"     "Hungary"    
#> [111] "Hungary"     "Hungary"     "Iceland"     "Iceland"     "Iceland"    
#> [116] "Iceland"     "Iceland"     "Iceland"     "Iceland"     "Iceland"    
#> [121] "Italy"       "Italy"       "Italy"       "Italy"       "Italy"      
#> [126] "Italy"       "Italy"       "Italy"       "Lithuania"   "Lithuania"  
#> [131] "Lithuania"   "Lithuania"   "Lithuania"   "Lithuania"   "Lithuania"  
#> [136] "Lithuania"   "Luxembourg"  "Luxembourg"  "Luxembourg"  "Luxembourg" 
#> [141] "Luxembourg"  "Luxembourg"  "Luxembourg"  "Luxembourg"  "Latvia"     
#> [146] "Latvia"      "Latvia"      "Latvia"      "Latvia"      "Latvia"     
#> [151] "Latvia"      "Latvia"      "Montenegro"  "Montenegro"  "Montenegro" 
#> [156] "Montenegro"  "Montenegro"  "Montenegro"  "Montenegro"  "Montenegro" 
#> [161] "Netherlands" "Netherlands" "Netherlands" "Netherlands" "Netherlands"
#> [166] "Netherlands" "Netherlands" "Netherlands" "Norway"      "Norway"     
#> [171] "Norway"      "Norway"      "Norway"      "Norway"      "Norway"     
#> [176] "Norway"      "Poland"      "Poland"      "Poland"      "Poland"     
#> [181] "Poland"      "Poland"      "Poland"      "Poland"      "Portugal"   
#> [186] "Portugal"    "Portugal"    "Portugal"    "Portugal"    "Portugal"   
#> [191] "Portugal"    "Portugal"    "Romania"     "Romania"     "Romania"    
#> [196] "Romania"     "Romania"     "Romania"     "Romania"     "Romania"    
#> [201] "Slovakia"    "Slovakia"    "Slovakia"    "Slovakia"    "Slovakia"   
#> [206] "Slovakia"    "Slovakia"    "Slovakia"    "Slovenia"    "Slovenia"   
#> [211] "Slovenia"    "Slovenia"    "Slovenia"    "Slovenia"    "Slovenia"   
#> [216] "Slovenia"    "Turkey"      "Turkey"      "Turkey"      "Turkey"     
#> [221] "Turkey"      "Turkey"      "Turkey"      "Turkey"      "US"         
#> [226] "US"          "US"          "US"          "US"          "US"         
#> [231] "US"          "US"